+++ frontend/components/Window/WindowHeader.js
