+++ frontend/components/WindowModules/Search.js
@@ -1,11 +1,9 @@
-// Teleoscope.js
 import React, { useState } from 'react';
-import { useCookies } from "react-cookie";
+import PropTypes from 'prop-types';
 
 // mui
 import { styled, alpha } from '@mui/material/styles';
 import InputBase from '@mui/material/InputBase';
-
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import CssBaseline from '@mui/material/CssBaseline';
@@ -24,19 +22,22 @@ import AddIcon from '@mui/icons-material/Add';
 import SearchIcon from '@mui/icons-material/Search';
 import MoreIcon from '@mui/icons-material/MoreVert';
 import LoadingButton from '@mui/lab/LoadingButton';
+import useScrollTrigger from '@mui/material/useScrollTrigger';
+import Slide from '@mui/material/Slide';
+import Container from '@mui/material/Container';
 
 // actions
 import { useSelector, useDispatch } from "react-redux";
-
+import { updateWindow } from "../../actions/windows";
 
 // custom components
-import PostList from "./PostList"
-import CloseButton from "./CloseButton"
-import { WindowHeader } from './WindowHeader';
-import { WindowBody } from './WindowBody';
+import PostList from "../PostList"
+import CloseButton from "../CloseButton"
+import { WindowHeader } from '../Window/WindowHeader';
+import { WindowBody } from '../Window/WindowBody';
 
 // util
-import useSWRAbstract from "../util/swr"
+import useSWRAbstract from "../../util/swr"
 
 const Search = styled('div')(({ theme }) => ({
   position: 'relative',
@@ -79,60 +80,89 @@ const StyledInputBase = styled(InputBase)(({ theme }) => ({
   },
 }));
 
+function HideOnScroll(props) {
+  const { children, window } = props;
+  // Note that you normally won't need to set the window ref as useScrollTrigger
+  // will default to window.
+  // This is only being set here because the demo is in an iframe.
+  const trigger = useScrollTrigger({
+    target: window ? window() : undefined,
+  });
 
-// abstracting for Window Header & Body
-const TeleoscopeHeader = ({props}) => {
-  const teleoscope_id = props.id.split("_")[0];
-  const [cookies, setCookie] = useCookies(["user"]);
-  const { user } = useSWRAbstract("user", `/api/users/${cookies.user}`);
-  const { teleoscope, teleoscope_loading } = useSWRAbstract("teleoscope", `/api/teleoscopes/${teleoscope_id}`);
+  return (
+    <Slide appear={false} direction="down" in={!trigger}>
+      {children}
+    </Slide>
+  );
+}
 
-  var data = [];
-  if (teleoscope) {
-    var history = teleoscope["history"];
-    var history_item = history[history.length - 1];
-    data = history_item["rank_slice"];
-  }
+HideOnScroll.propTypes = {
+  children: PropTypes.element.isRequired,
+  /**
+   * Injected by the documentation to work in an iframe.
+   * You won't need it on your project.
+   */
+  window: PropTypes.func,
+};
+
+let globalQuery = "";
+
+
+// splits the search into the header and the body
+const SearchHeader = ({ props }) => {
+
+const [query, setQuery] = useState("");
+const dispatch = useDispatch();
+
+const handleSetQuery = (e) => {
+  setTimeout(() => {
+    setQuery(e.target.value);
+    globalQuery = e.target.value;
+    dispatch(updateWindow({ i: "%search", term: e.target.value }));
+  }, 1000);
+};
 
   return (
     <div>
+      <Search>
+        <SearchIconWrapper>
+          <SearchIcon />
+        </SearchIconWrapper>
+        <StyledInputBase
+          placeholder="Search…"
+          inputProps={{ 'aria-label': 'search' }}
+          onChange={(e) => handleSetQuery(e)}/>
+      </Search>
       <Typography variant="h5" gutterBottom component="div" sx={{ p: 2, pb: 0 }}>
-        Teleoscope: {teleoscope?.history[teleoscope?.history.length - 1].label}
+        Search {query != "" ? `"${query}"` : "all posts"}
       </Typography>
     </div>
-  )
-
+  );
 }
 
-const TeleoscopeBody = ({props}) => {
-  const teleoscope_id = props.id.split("_")[0];
-  const [cookies, setCookie] = useCookies(["user"]);
-  const { user } = useSWRAbstract("user", `/api/users/${cookies.user}`);
-  const { teleoscope, teleoscope_loading } = useSWRAbstract("teleoscope", `/api/teleoscopes/${teleoscope_id}`);
+const SearchBody = ({ props }) => {
+  const { posts, posts_loading, posts_error } = useSWRAbstract("posts", `/api/cleanposts/${globalQuery}`);
 
-  var data = [];
-  if (teleoscope) {
-    var history = teleoscope["history"];
-    var history_item = history[history.length - 1];
-    data = history_item["rank_slice"];
-  }
+  // this is a hard-coded hack for ranking of post_id
+  const data = posts ? posts.map((post) => { return [post.id, 1.0]; }) : [];
 
   return (
     <div>
-      {teleoscope_loading ? <LoadingButton loading={true} /> : <PostList pagination={true} data={data}></PostList>}
+      {posts_loading ? <LoadingButton loading={true} /> : <PostList pagination={true} data={data}></PostList>}
     </div>
   )
 }
 
-export default function Teleoscope(props) {
+
+
+export default function BottomAppBar(props) {
   return (
     <>
       <WindowHeader>
-        <TeleoscopeHeader props={props} />
+        <SearchHeader props={props} />
       </WindowHeader>
-
       <WindowBody>
-        <TeleoscopeBody props={props} />
+        <SearchBody props={props} />
       </WindowBody>
     </>
   );
