+++ frontend/components/WindowModules/Teleoscope.js
@@ -1,9 +1,11 @@
+// Teleoscope.js
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+import { useCookies } from "react-cookie";
 
 // mui
 import { styled, alpha } from '@mui/material/styles';
 import InputBase from '@mui/material/InputBase';
+
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import CssBaseline from '@mui/material/CssBaseline';
@@ -22,22 +24,19 @@ import AddIcon from '@mui/icons-material/Add';
 import SearchIcon from '@mui/icons-material/Search';
 import MoreIcon from '@mui/icons-material/MoreVert';
 import LoadingButton from '@mui/lab/LoadingButton';
-import useScrollTrigger from '@mui/material/useScrollTrigger';
-import Slide from '@mui/material/Slide';
-import Container from '@mui/material/Container';
 
 // actions
 import { useSelector, useDispatch } from "react-redux";
-import { updateWindow } from "../actions/windows";
+
 
 // custom components
-import PostList from "./PostList"
-import CloseButton from "./CloseButton"
-import { WindowHeader } from './WindowHeader';
-import { WindowBody } from './WindowBody';
+import PostList from "../PostList"
+import CloseButton from "../CloseButton"
+import { WindowHeader } from '../Window/WindowHeader';
+import { WindowBody } from '../Window/WindowBody';
 
 // util
-import useSWRAbstract from "../util/swr"
+import useSWRAbstract from "../../util/swr"
 
 const Search = styled('div')(({ theme }) => ({
   position: 'relative',
@@ -80,89 +79,60 @@ const StyledInputBase = styled(InputBase)(({ theme }) => ({
   },
 }));
 
-function HideOnScroll(props) {
-  const { children, window } = props;
-  // Note that you normally won't need to set the window ref as useScrollTrigger
-  // will default to window.
-  // This is only being set here because the demo is in an iframe.
-  const trigger = useScrollTrigger({
-    target: window ? window() : undefined,
-  });
 
-  return (
-    <Slide appear={false} direction="down" in={!trigger}>
-      {children}
-    </Slide>
-  );
-}
+// abstracting for Window Header & Body
+const TeleoscopeHeader = ({props}) => {
+  const teleoscope_id = props.id.split("_")[0];
+  const [cookies, setCookie] = useCookies(["user"]);
+  const { user } = useSWRAbstract("user", `/api/users/${cookies.user}`);
+  const { teleoscope, teleoscope_loading } = useSWRAbstract("teleoscope", `/api/teleoscopes/${teleoscope_id}`);
 
-HideOnScroll.propTypes = {
-  children: PropTypes.element.isRequired,
-  /**
-   * Injected by the documentation to work in an iframe.
-   * You won't need it on your project.
-   */
-  window: PropTypes.func,
-};
-
-let globalQuery = "";
-
-
-// splits the search into the header and the body
-const SearchHeader = ({ props }) => {
-
-const [query, setQuery] = useState("");
-const dispatch = useDispatch();
-
-const handleSetQuery = (e) => {
-  setTimeout(() => {
-    setQuery(e.target.value);
-    dispatch(updateWindow({ i: "%search", term: e.target.value }));
-  }, 1000);
-};
+  var data = [];
+  if (teleoscope) {
+    var history = teleoscope["history"];
+    var history_item = history[history.length - 1];
+    data = history_item["rank_slice"];
+  }
 
   return (
     <div>
-      <Search>
-        <SearchIconWrapper>
-          <SearchIcon />
-        </SearchIconWrapper>
-        <StyledInputBase
-          placeholder="Search…"
-          inputProps={{ 'aria-label': 'search' }}
-          onChange={(e) => handleSetQuery(e)} />
-      </Search>
-      {globalQuery = query}
       <Typography variant="h5" gutterBottom component="div" sx={{ p: 2, pb: 0 }}>
-        Search {query != "" ? `"${query}"` : "all posts"}
+        Teleoscope: {teleoscope?.history[teleoscope?.history.length - 1].label}
       </Typography>
     </div>
-  );
+  )
+
 }
 
-const SearchBody = ({ props }) => {
-  const { posts, posts_loading, posts_error } = useSWRAbstract("posts", `/api/cleanposts/${globalQuery}`);
+const TeleoscopeBody = ({props}) => {
+  const teleoscope_id = props.id.split("_")[0];
+  const [cookies, setCookie] = useCookies(["user"]);
+  const { user } = useSWRAbstract("user", `/api/users/${cookies.user}`);
+  const { teleoscope, teleoscope_loading } = useSWRAbstract("teleoscope", `/api/teleoscopes/${teleoscope_id}`);
 
-  // this is a hard-coded hack for ranking of post_id
-  const data = posts ? posts.map((post) => { return [post.id, 1.0]; }) : [];
+  var data = [];
+  if (teleoscope) {
+    var history = teleoscope["history"];
+    var history_item = history[history.length - 1];
+    data = history_item["rank_slice"];
+  }
 
   return (
     <div>
-      {posts_loading ? <LoadingButton loading={true} /> : <PostList pagination={true} data={data}></PostList>}
+      {teleoscope_loading ? <LoadingButton loading={true} /> : <PostList pagination={true} data={data}></PostList>}
     </div>
   )
 }
 
-
-
-export default function BottomAppBar(props) {
+export default function Teleoscope(props) {
   return (
     <>
       <WindowHeader>
-        <SearchHeader props={props} />
+        <TeleoscopeHeader props={props} />
       </WindowHeader>
+
       <WindowBody>
-        <SearchBody props={props} />
+        <TeleoscopeBody props={props} />
       </WindowBody>
     </>
   );
